Report rules in sorted order
Resolves #39.
 
The fix involves sorting the rules by name and then using this order to iterate over the rules and their associated results. The sorting for the results is already taken care of by `ByRule()`.

I didn't add any tests because the addition is comparable to the sorting being done in `ByRule()` for which there are no tests. 

Please let me know if you have any questions or any changes to suggest.